# Joomla String Class Optimisation