--- a/libraries/legacy/application/application.php
+++ b/libraries/legacy/application/application.php
@@ -389,9 +389,7 @@ class JApplication extends JApplicationBase
 		{
 			$document = JFactory::getDocument();
 
-			jimport('phputf8.utils.ascii');
-
-			if (($this->client->engine == JApplicationWebClient::TRIDENT) && !utf8_is_ascii($url))
+			if (($this->client->engine == JApplicationWebClient::TRIDENT) && !JString::is_ascii($url))
 			{
 				// MSIE type browser and/or server cause issues when url contains utf8 character,so use a javascript redirect method
 				echo '<html><head><meta http-equiv="content-type" content="text/html; charset=' . $document->getCharset() . '" />'