… libraries/vendor/joomla/string/src/String.php