@wilsonge 2015-04-16
Hi Thanks for this PR! Anything that touches the framework files should be made to the string package in the framework repository (https://github.com/joomla-framework/string) If you're happy to do that it would be really appreciated :) Thanks!