@orware 2015-04-16
Thanks George,

That was going to be my original plan once I noticed it was one of the framework packages...however the current version of the package we're using in Joomla doesn't seem to match up with the Master String branch in the framework repository (as an example, in the CMS there's just a String.php file, but in the framework String.php defers to another file called StringHelper.php which actually contains the function definitions...since I didn't know when/if those updates would be brought into the CMS I thought it would be better to create my PR against what is currently in the CMS first).

Maybe you have some extra info about whether or not the String package is going to be updated in the CMS soon? If it will be, I can certainly create an additional PR against the framework code.