@wilsonge 2015-04-16
We will be merging the string updates into 3.5 - as there is a low chance of a b/c break because of the class rename (PHP 7 forbid us to use the word string as it's now protected). https://github.com/joomla/joomla-cms/pull/6600 The PR is already in for it though