@orware 2015-04-16
Thanks for the update...I'll work on the PR against the Framework package sometime today ;-).