@orware 2015-04-16
I'm going to go ahead and close this PR since Michael merged in the separate PR I created for the String package earlier today.